* cont.c (fiber_set_stack_location), iseq.c (prepare_iseq_build),   node.c (dump_node): suppress warnings.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@27808 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ iseq.c
@@ -229,7 +229,7 @@ prepare_iseq_build(rb_iseq_t *iseq,
     iseq->name = name;
     iseq->filename = filename;
     iseq->filepath = filepath == Qnil ? Qnil : rb_realpath_internal(Qnil, filepath, 1);
-    iseq->line_no = line_no;
+    iseq->line_no = (unsigned short)line_no; /* TODO: really enough? */
     iseq->defined_method_id = 0;
     iseq->mark_ary = rb_ary_tmp_new(3);
     OBJ_UNTRUST(iseq->mark_ary);
